+++ Cesium3DTilesSelection/src/TileUtilities.cpp
@@ -1,5 +1,6 @@
 #include "TileUtilities.h"
 
+#include <CesiumGeometry/IntersectionTests.h>
 #include <CesiumGeospatial/BoundingRegion.h>
 #include <CesiumGeospatial/BoundingRegionWithLooseFittingHeights.h>
 #include <CesiumGeospatial/GlobeRectangle.h>
@@ -7,6 +8,7 @@
 #include <variant>
 
 using namespace CesiumGeospatial;
+using namespace CesiumGeometry;
 
 namespace Cesium3DTilesSelection {
 namespace CesiumImpl {