+++ Cesium3DTilesSelection/src/TileUtilities.h
@@ -2,6 +2,7 @@
 
 #include "Cesium3DTilesSelection/BoundingVolume.h"
 
+#include <CesiumGeometry/Ray.h>
 #include <CesiumGeospatial/CartographicPolygon.h>
 #include <CesiumGeospatial/GlobeRectangle.h>
 
@@ -32,5 +33,6 @@ bool outsidePolygons(
     const BoundingVolume& boundingVolume,
     const std::vector<CesiumGeospatial::CartographicPolygon>&
         cartographicPolygons) noexcept;
+
 } // namespace CesiumImpl
 } // namespace Cesium3DTilesSelection