+++ CesiumGltfContent/src/GltfUtilities.cpp
@@ -1,4 +1,6 @@
 #include <CesiumGeometry/Axis.h>
+#include <CesiumGeometry/IntersectionTests.h>
+#include <CesiumGeometry/Ray.h>
 #include <CesiumGeometry/Transforms.h>
 #include <CesiumGeospatial/BoundingRegionBuilder.h>
 #include <CesiumGltf/AccessorView.h>
@@ -6,6 +8,7 @@
 #include <CesiumGltfContent/GltfUtilities.h>
 #include <CesiumGltfContent/SkirtMeshMetadata.h>
 
+#include <limits>
 #include <vector>
 
 using namespace CesiumGltf;
@@ -211,6 +214,192 @@ GltfUtilities::parseGltfCopyright(const CesiumGltf::Model& gltf) {
     bufferView.buffer = int32_t(destinationIndex);
     bufferView.byteOffset += int64_t(start);
   }
+}
+
+namespace {
+
+double signAwareMin(double n1, double n2) {
+  if (n2 < n1)
+    std::swap(n1, n2);
+  return n1 < 0 ? n2 : n1;
+}
+
+template <class T>
+bool intersectRayPrimitiveParametric(
+    const CesiumGeometry::Ray& ray,
+    const CesiumGltf::Model& model,
+    const CesiumGltf::MeshPrimitive& primitive,
+    bool cullBackFaces,
+    double& tMin) {
+
+  auto positionAccessorIt = primitive.attributes.find("POSITION");
+  if (positionAccessorIt == primitive.attributes.end()) {
+    return false;
+  }
+
+  int positionAccessorID = positionAccessorIt->second;
+  const Accessor* pPositionAccessor =
+      Model::getSafe(&model.accessors, positionAccessorID);
+  if (!pPositionAccessor) {
+    return false;
+  }
+
+  const std::vector<double>& min = pPositionAccessor->min;
+  const std::vector<double>& max = pPositionAccessor->max;
+
+  double t;
+  if (!CesiumGeometry::IntersectionTests::rayAABBParametric(
+          ray,
+          CesiumGeometry::AxisAlignedBox(
+              min[0],
+              min[1],
+              min[2],
+              max[0],
+              max[1],
+              max[2]),
+          t)) {
+    return false;
+  }
+
+  AccessorView<T> indicesView(model, primitive.indices);
+  AccessorView<glm::vec3> positionView(model, *pPositionAccessor);
+
+  tMin = -std::numeric_limits<double>::max();
+  double tCurr;
+  bool intersected = false;
+
+  if (primitive.mode == CesiumGltf::MeshPrimitive::Mode::TRIANGLES) {
+    for (int32_t i = 0; i < indicesView.size(); i += 3) {
+      if (CesiumGeometry::IntersectionTests::rayTriangleParametric(
+              ray,
+              glm::dvec3(positionView[static_cast<int32_t>(indicesView[i])]),
+              glm::dvec3(
+                  positionView[static_cast<int32_t>(indicesView[i + 1])]),
+              glm::dvec3(
+                  positionView[static_cast<int32_t>(indicesView[i + 2])]),
+              tCurr,
+              cullBackFaces)) {
+        intersected = true;
+        tMin = signAwareMin(tMin, tCurr);
+      }
+    }
+  } else {
+    for (int32_t i = 0; i < indicesView.size() - 2; ++i) {
+      if (i % 2) {
+        if (CesiumGeometry::IntersectionTests::rayTriangleParametric(
+                ray,
+                glm::dvec3(positionView[static_cast<int32_t>(indicesView[i])]),
+                glm::dvec3(
+                    positionView[static_cast<int32_t>(indicesView[i + 2])]),
+                glm::dvec3(
+                    positionView[static_cast<int32_t>(indicesView[i + 1])]),
+                tCurr,
+                true)) {
+          intersected = true;
+          tMin = signAwareMin(tMin, tCurr);
+        }
+      } else {
+        if (CesiumGeometry::IntersectionTests::rayTriangleParametric(
+                ray,
+                glm::dvec3(positionView[static_cast<int32_t>(indicesView[i])]),
+                glm::dvec3(
+                    positionView[static_cast<int32_t>(indicesView[i + 1])]),
+                glm::dvec3(
+                    positionView[static_cast<int32_t>(indicesView[i + 2])]),
+                tCurr,
+                true)) {
+          intersected = true;
+          tMin = signAwareMin(tMin, tCurr);
+        }
+      }
+    }
+  }
+  return intersected;
+}
+} // namespace
+
+bool GltfUtilities::intersectRayGltfModelParametric(
+    const CesiumGeometry::Ray& ray,
+    const CesiumGltf::Model& gltf,
+    double& tMin,
+    bool cullBackFaces,
+    const glm::dmat4x4& modelToWorld) {
+  glm::dmat4x4 rootTransform = applyRtcCenter(gltf, modelToWorld);
+  rootTransform = applyGltfUpAxisTransform(gltf, rootTransform);
+
+  tMin = -std::numeric_limits<double>::max();
+  bool intersected = false;
+
+  gltf.forEachPrimitiveInScene(
+      -1,
+      [ray, cullBackFaces, rootTransform, &intersected, &tMin](
+          const CesiumGltf::Model& model,
+          const CesiumGltf::Node& /*node*/,
+          const CesiumGltf::Mesh& /*mesh*/,
+          const CesiumGltf::MeshPrimitive& primitive,
+          const glm::dmat4& nodeTransform) {
+        if (primitive.mode != MeshPrimitive::Mode::TRIANGLES &&
+            primitive.mode != MeshPrimitive::Mode::TRIANGLE_STRIP) {
+          return;
+        }
+
+        glm::dmat4x4 worldToPrimitive =
+            glm::inverse(rootTransform * nodeTransform);
+
+        double tCurr = -std::numeric_limits<double>::max();
+        bool intersectedPrimitive = false;
+
+        switch (model.accessors[static_cast<uint32_t>(primitive.indices)]
+                    .componentType) {
+        case Accessor::ComponentType::UNSIGNED_BYTE:
+          intersectedPrimitive = intersectRayPrimitiveParametric<uint8_t>(
+              ray.transform(worldToPrimitive),
+              model,
+              primitive,
+              cullBackFaces,
+              tCurr);
+          break;
+        case Accessor::ComponentType::UNSIGNED_SHORT:
+          intersectedPrimitive = intersectRayPrimitiveParametric<uint16_t>(
+              ray.transform(worldToPrimitive),
+              model,
+              primitive,
+              cullBackFaces,
+              tCurr);
+          break;
+        case Accessor::ComponentType::UNSIGNED_INT:
+          intersectedPrimitive = intersectRayPrimitiveParametric<uint32_t>(
+              ray.transform(worldToPrimitive),
+              model,
+              primitive,
+              cullBackFaces,
+              tCurr);
+          break;
+        }
+        if (intersectedPrimitive) {
+          intersected = true;
+          tMin = signAwareMin(tMin, tCurr);
+        }
+      });
+
+  return intersected;
+}
+
+std::optional<glm::dvec3> GltfUtilities::intersectRayGltfModel(
+    const CesiumGeometry::Ray& ray,
+    const CesiumGltf::Model& gltf,
+    bool cullBackFaces,
+    const glm::dmat4x4& modelToWorld) {
+  double t;
+  return intersectRayGltfModelParametric(
+             ray,
+             gltf,
+             t,
+             cullBackFaces,
+             modelToWorld) &&
+                 t >= 0
+             ? std::make_optional<glm::dvec3>(ray.getPointAlongRay(t))
+             : std::nullopt;
 }
 
 } // namespace CesiumGltfContent